refactor(tasks): move `collectBaseClasses` task registration to extension function
